+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog
 - Fix #129: Snippet visibility selection to "Public" broken
 - Fix #130: Menu item of space page not activated
 - Fix #124: one- and two column layout broken on space page
+- Fix: Snippet notification content description missing
 
 1.0.8 (May 29, 2020)
 --------------------